+++ Domain/Exceptions/RecipeNotFoundException.cs
@@ -2,7 +2,7 @@ namespace Domain.Exceptions;
 
 public class RecipeNotFoundException : Exception
 {
-    public RecipeNotFoundException(int recipeId)
+    public RecipeNotFoundException(long recipeId)
     : base($"Recipe with id {recipeId} was not found.")
     {
     }
