+++ Domain/Repositories/IRecipeRepository.cs
@@ -6,6 +6,6 @@ public interface IRecipeRepository
 {
     void Add(Recipe recipe);
     void Remove(Recipe recipe);
-    Task<Recipe?> GetByIdAsync(int id);
+    Task<Recipe?> GetByIdAsync(long id);
     Task<List<Recipe>> GetSummaries();
 }
